backend/tests/database_test.py
"""
Module for testing database connection and Vault integration.
"""

import unittest
from unittest.mock import patch, MagicMock
from backend import database  # Adjust the import path to match your project structure

class TestDatabase(unittest.TestCase):
    """
    Test case for the database module.
    """

    @patch('backend.database.client')
    def test_login_with_userpass_success(self, mock_client):
        """
        Test successful login with userpass.
        """
        mock_login_response = {'auth': {'client_token': 'test-token'}}
        mock_client.auth.userpass.login.return_value = mock_login_response

        token = database.login_with_userpass('testuser', 'testpass')
        self.assertEqual(token, 'test-token')
        self.assertEqual(database.client.token, 'test-token')

    @patch('backend.database.client')
    def test_login_with_userpass_failure(self, mock_client):
        """
        Test login failure with userpass.
        """
        mock_client.auth.userpass.login.side_effect = Exception("Login failed")
        with self.assertRaises(Exception) as context:
            database.login_with_userpass('testuser', 'testpass')
        self.assertTrue('Login failed' in str(context.exception))

    @patch('backend.database.client')
    def test_login_with_userpass_invalid_credentials(self, mock_client):
        """
        Test login with invalid credentials.
        """
        mock_client.auth.userpass.login.return_value = {'auth': {'client_token': 'test-token'}}
        # Assuming the function handles invalid credentials by returning None or a specific value
        token = database.login_with_userpass('invaliduser', 'invalidpass')
        self.assertIn(token, [None, 'test-token'])  # Adjust based on your function's behavior

    @patch('backend.database.client')
    def test_read_secret_success(self, mock_client):
        """
        Test successful secret reading.
        """
        mock_read_response = {
            'data': {'data': {'username': 'testuser', 'password': 'testpass'}}
        }
        mock_client.secrets.kv.v2.read_secret_version.return_value = mock_read_response

        secret_data = database.read_secret(mock_client, 'credentials', 'db')
        self.assertEqual(secret_data, {'username': 'testuser', 'password': 'testpass'})

    @patch('backend.database.client')
    def test_read_secret_failure(self, mock_client):
        """
        Test secret reading failure.
        """
        mock_client.secrets.kv.v2.read_secret_version.side_effect = Exception("Read failed")

        secret_data = database.read_secret(mock_client, 'credentials', 'db')
        self.assertIsNone(secret_data)

    @patch('backend.database.client')
    def test_read_secret_empty_response(self, mock_client):
        """
        Test reading a secret with an empty response.
        """
        mock_client.secrets.kv.v2.read_secret_version.return_value = {'data': {'data': None}}
        secret = database.read_secret(mock_client, 'path/to/secret')
        self.assertIsNone(secret)

    @patch('backend.database.create_engine')
    @patch('backend.database.sessionmaker')
    def test_create_database_session(self, mock_sessionmaker, mock_create_engine):
        """
        Test successful database session creation.
        """
        mock_engine = MagicMock()
        mock_create_engine.return_value = mock_engine

        mock_sessionmaker_instance = MagicMock()
        mock_sessionmaker.return_value = mock_sessionmaker_instance

        database_url = 'mysql+pymysql://user:pass@localhost/db'
        engine, session_local = database.create_database_session(database_url)

        mock_create_engine.assert_called_once_with(database_url, pool_pre_ping=True)
        self.assertEqual(engine, mock_engine)
        self.assertEqual(session_local, mock_sessionmaker_instance)

    @patch('backend.database.client')
    def test_read_secret_failure(self, mock_client):
        """
        Test failure in reading a secret.
        """
        mock_client.secrets.kv.v2.read_secret_version.side_effect = Exception("Read failed")
        with self.assertRaises(Exception) as context:
            database.read_secret(mock_client, 'path/to/secret')
        self.assertTrue('Read failed' in str(context.exception))

if __name__ == "__main__":
    unittest.main()
